Add files via upload
#UDEMY
Q. 1) What are all different subjects for which Udemy is offering courses ? 
Q. 2) Which subject has the maximum number of courses. 
Q. 3) Show all the courses which are Free of Cost. 
Q. 4) Show all the courses which are Paid. 
Q. 5) Which are Top Selling Courses ? 
Q. 6) Which are Least Selling Courses ? 
Q. 7) Show all courses of Graphic Design where the price is below 100 ? 
Q. 8) List out all the courses that are related to 'Python'. 
Q. 9) What are courses that were published in the year 2015 ? 
Q. 10) What is the Max. Number of Subscribers for Each Level of courses ?